requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart
pillow
torch
torchvision
dlib
numpy
opencv-python
